rename bookitem and bookcontainer
